Use runWidget; remove runAppWithoutImplicitView (#20)

+++ lib/raw_static.dart
@@ -31,7 +31,7 @@ void _handleOnDrawFrame() {
 
   for (final FlutterView view in PlatformDispatcher.instance.views) {
     final Size logicalSize = view.physicalSize / view.devicePixelRatio;
-    final Color color = _colors[(view.viewId as int) % _colors.length];
+    final Color color = _colors[view.viewId % _colors.length];
 
     final ParagraphBuilder paragraphBuilder = ParagraphBuilder(
       ParagraphStyle(textDirection: TextDirection.ltr, fontSize: 24),
